+++ src/fidesops/ops/schemas/saas/saas_config.py
@@ -222,6 +222,7 @@ class SaaSType(Enum):
     zendesk = "zendesk"
     custom = "custom"
     sendgrid = "sendgrid"
+    datadog = "datadog"
 
 
 class SaaSConfigBase(BaseModel):